build(scripts): refactor clear-container.sh

+++ scripts/clean-local.sh
@@ -1,6 +1,11 @@
-echo '# stop and remove running container'
-docker stop $(docker ps -q)
-docker rm $(docker ps -aq)
+if [ "$(docker ps -q)" ]; then
+    echo '# stop running container'
+    docker stop $(docker ps -q)
+fi
+if [ "$(docker ps -aq)" ]; then
+    echo '# remove exited container'
+    docker rm $(docker ps -aq)
+fi
 docker-compose down
 echo '# removing backend dependencies'
 rm -rf ./node_modules
