+++ scripts/clear-container.sh
@@ -1,3 +1,8 @@
-echo '# stop and remove running container'
-docker stop $(docker ps -q)
-docker rm $(docker ps -aq)
+if [ "$(docker ps -q)" ]; then
+    echo '# stop running container'
+    docker stop $(docker ps -q)
+fi
+if [ "$(docker ps -aq)" ]; then
+    echo '# remove exited container'
+    docker rm $(docker ps -aq)
+fi
